Bookmark utils. Split server routes into own route files
