+++ test-app/routes/charts.js
@@ -1,0 +1,54 @@
+const mockBarChartData = require('../mocks/mockChartData/mockBarChartData')
+const mockPieChartData = require('../mocks/mockChartData/mockPieChartData')
+const mockLineChartData = require('../mocks/mockChartData/mockLineChartData')
+const mockMulitChartData = require('../mocks/mockChartData/mockMultiChartData')
+const mockScoreCards = require('../mocks/mockScoreCards/mockScorecards')
+
+const chartsRoutes = (app) => {
+  app.get('/charts', (req, res) => {
+    res.render('charts.njk', {
+      title: 'Charts',
+      barCharts: mockBarChartData,
+      pieCharts: mockPieChartData,
+      lineCharts: mockLineChartData,
+      multiCharts: mockMulitChartData,
+    })
+  })
+
+  app.get('/charts/bar', (req, res) => {
+    res.render('charts.njk', {
+      title: 'Charts',
+      barCharts: mockBarChartData,
+    })
+  })
+
+  app.get('/charts/donut', (req, res) => {
+    res.render('charts.njk', {
+      title: 'Charts',
+      pieCharts: mockPieChartData,
+    })
+  })
+
+  app.get('/charts/line', (req, res) => {
+    res.render('charts.njk', {
+      title: 'Charts',
+      lineCharts: mockLineChartData,
+    })
+  })
+
+  app.get('/charts/multi', (req, res) => {
+    res.render('charts.njk', {
+      title: 'Charts',
+      multiCharts: mockMulitChartData,
+    })
+  })
+
+  app.get('/scorecards', (req, res) => {
+    res.render('scorecards.njk', {
+      title: 'Score Cards',
+      scorecards: mockScoreCards,
+    })
+  })
+}
+
+module.exports = chartsRoutes
